requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
gitdb==4.0.10
GitPython==3.1.37
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.26.0
packaging==23.2
pandas==2.1.1
pip==22.0.3
Pillow==10.0.1
plotly==5.15.0
plotly-express==0.4.1
protobuf==4.24.4
pyarrow==13.0.0
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
six==1.16.0
smmap==5.0.1
streamlit==1.27.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
validators==0.22.0
zipp==3.17.0